Isolate YAML config code to `src/config`
