+++ config/default_configs/default_config.yml
@@ -103,7 +103,7 @@ idealized_h2o:
   help: "Use idealized H2O in radiation model [`false` (default), `true`]"
   value: false
 dt_rad:
-  help: "Time between calling radiation callback for sphere configurations"
+  help: "Time between calling radiation callback for sphere configurations. Only applies to RRTMGP modes, ignored for `held_suarez`, which is applied every stage."
   value: "6hours"
 idealized_clouds:
   help: "Use idealized clouds in radiation model [`false` (default), `true`]"
@@ -190,9 +190,6 @@ start_date:
   value: "20100101"
 era5_initial_condition_dir:
   help: "Directory containing ERA5 initial condition files. Filenames inferred from start_date [none (default)]. Generated with `https://github.com/CliMA/WeatherQuest`."
-  value: ~
-forcing:
-  help: "Forcing [`nothing` (default), `held_suarez`]"
   value: ~
 test_dycore_consistency:
   help: "Test dycore consistency [`false` (default), `true`]"
