+++ config/longrun_configs/longrun_dry_held_suarez.yml
@@ -1,7 +1,7 @@
 implicit_diffusion: false
 approximate_linear_solve_iters: 1
 t_end: "360days"
-forcing: "held_suarez"
+rad: "held_suarez"
 dt_save_state_to_disk: "10days"
 disable_surface_flux_tendency: true
 toml: [toml/longrun_held_suarez.toml]
