+++ config/longrun_configs/longrun_moist_held_suarez.yml
@@ -3,6 +3,6 @@ approximate_linear_solve_iters: 1
 t_end: "360days"
 microphysics_model: "0M"
 vert_diff: "VerticalDiffusion"
-forcing: "held_suarez"
+rad: "held_suarez"
 dt_save_state_to_disk: "10days"
 toml: [toml/longrun_held_suarez.toml]
