+++ perf/tabulate_perf_summary.jl
@@ -1,7 +1,7 @@
 import Dates
 import JSON
 import OrderedCollections
-include(joinpath(@__DIR__, "..", "src", "solver", "yaml_helper.jl"))
+include(joinpath(@__DIR__, "..", "src", "config", "yaml_helper.jl"))
 
 function sorted_dataset_folder(; dir = pwd())
     matching_paths = String[]
