+++ src/callbacks/get_callbacks.jl
@@ -176,7 +176,7 @@ function parse_frequency_to_schedule(
     else
         period_seconds = FT(time_to_seconds(frequency_str))
         period_dates =
-            CA.promote_period.(Dates.Second(period_seconds))
+            promote_period.(Dates.Second(period_seconds))
     end
 
     return CAD.EveryCalendarDtSchedule(
@@ -216,9 +216,9 @@ function validate_checkpoint_diagnostics_consistency(
     periods_reductions,
 )
     if checkpoint_frequency != Inf
-        if any(x -> !CA.isdivisible(checkpoint_frequency, x), periods_reductions)
-            accum_str = join(CA.promote_period.(collect(periods_reductions)), ", ")
-            checkpt_str = CA.promote_period(checkpoint_frequency)
+        if any(x -> !isdivisible(checkpoint_frequency, x), periods_reductions)
+            accum_str = join(promote_period.(collect(periods_reductions)), ", ")
+            checkpt_str = promote_period(checkpoint_frequency)
             @warn """The checkpointing frequency \
             (checkpoint_frequency = $checkpt_str) should be an integer \
             multiple of all diagnostics accumulation periods ($accum_str) \
@@ -337,7 +337,7 @@ function scheduled_callback(
 
     if !isnothing(checkpoint_frequency) && checkpoint_frequency != Inf
         dt_s = Dates.Second(round(Int, dt_seconds_val))
-        if !CA.isdivisible(checkpoint_frequency, dt_s)
+        if !isdivisible(checkpoint_frequency, dt_s)
             @warn "$(nameof(affect!)) period ($dt_s) is not an even divisor of the checkpoint frequency ($checkpoint_frequency). This simulation will not be reproducible when restarted."
         end
     end
@@ -353,6 +353,9 @@ function radiation_callback(
     t_end,
     checkpoint_frequency,
 )
+    # `dt_rad` only governs RRTMGP-style radiation. For `rad: held_suarez`
+    # (and any other non-RRTMGP mode), the forcing is folded into
+    # `remaining_tendency!` and applied every stage; `dt_rad` is ignored.
     radiation_mode isa RRTMGPI.AbstractRRTMGPMode || return ()
     return scheduled_callback(
         rrtmgp_model_callback!,
