+++ src/simulation/restart.jl
@@ -1,0 +1,171 @@
+import ClimaCore: InputOutput
+import ClimaUtilities.OutputPathGenerator
+import ClimaUtilities.TimeManager: ITime
+
+#####
+##### Restart file loading + output directory resolution.
+#####
+
+function get_state_restart(
+    restart_file,
+    start_date,
+    atmos_model_hash,
+    comms_ctx,
+)
+    @assert !isnothing(restart_file)
+    reader = InputOutput.HDF5Reader(restart_file, comms_ctx)
+    Y = InputOutput.read_field(reader, "Y")
+    # TODO: Do not use InputOutput.HDF5 directly
+    t_start = InputOutput.HDF5.read_attribute(reader.file, "time")
+    t_start = ITime(t_start; epoch = start_date)
+    if "atmos_model_hash" in keys(InputOutput.HDF5.attrs(reader.file))
+        atmos_model_hash_in_restart =
+            InputOutput.HDF5.read_attribute(reader.file, "atmos_model_hash")
+        if atmos_model_hash_in_restart != atmos_model_hash
+            @warn "Restart file $(restart_file) was constructed with a different AtmosModel"
+        end
+    end
+    return (Y, t_start)
+end
+
+"""
+    handle_restart(restart_file, t_start_original, start_date, model, context)
+
+Handle restart file loading with validation and logging.
+
+Validates that t_start is zero when restarting, loads state from restart file,
+logs restart information, and returns the state, t_start, and spaces.
+
+Returns:
+- `Y`: State loaded from restart file
+- `t_start`: Time from restart file (ITime)
+- `spaces`: Named tuple with center_space and face_space extracted from Y
+"""
+function handle_restart(
+    restart_file,
+    t_start_original,
+    start_date,
+    model,
+    context,
+)
+    # Validate t_start before restart (matches get_simulation behavior)
+    t_start_seconds =
+        t_start_original isa AbstractString ?
+        time_to_seconds(t_start_original) : Float64(t_start_original)
+    if t_start_seconds != 0
+        @warn "Non zero `t_start` passed with a restarting simulation. The provided `t_start` will be ignored."
+    end
+
+    (Y, t_start) = get_state_restart(
+        restart_file, start_date, hash(model), context,
+    )
+
+    @info "Restarting simulation from file" restart_file restart_time = string(t_start)
+
+    spaces = (; center_space = axes(Y.c), face_space = axes(Y.f))
+
+    return Y, t_start, spaces
+end
+
+auto_detect_restart_file(::OutputPathGenerator.OutputPathGeneratorStyle, _) =
+    error("auto_detect_restart_file works only with ActiveLink")
+
+"""
+    auto_detect_restart_file(::ActiveLinkStyle, base_output_dir)
+
+Return the most recent restart file in the directory structure in `base_output_dir`, if any.
+
+`auto_detect_restart_file` scans the content of `base_output_dir` matching the expected
+names for output folders generated by `ActiveLinkStyle` (e.g., `output_0000`, `output_0001`).
+It iterates through these folders sorted by number in descending order and returns the latest
+restart file (latest measured by the time in the file name) from the first folder that contains
+restart files matching the pattern `dayDDDD.SSSSS.hdf5`. This ensures that empty or incomplete
+higher-numbered folders are skipped in favor of folders that actually contain restart files.
+If no folder with restart files is found, return `nothing`: this means that the simulation
+cannot be automatically restarted.
+"""
+function auto_detect_restart_file(
+    ::OutputPathGenerator.ActiveLinkStyle,
+    base_output_dir,
+)
+    # if base_output_dir does not exist, we return restart_file = nothing because there is
+    # no restart file to be detected
+    isdir(base_output_dir) || return nothing
+
+    # output_dir will be something like ABC/DEF/output_1234
+    name_rx = r"output_(\d\d\d\d)"
+    restart_file_rx = r"day\d+\.\w+\.hdf5"
+    restart_file = nothing
+
+    existing_outputs =
+        filter(x -> !isnothing(match(name_rx, x)), readdir(base_output_dir))
+
+    isempty(existing_outputs) && return nothing
+
+    # Sort folders by number (descending) and find the first with restart files
+    for output_folder in sort(existing_outputs, rev = true)
+        folder_path = joinpath(base_output_dir, output_folder)
+        possible_restart_files =
+            filter(f -> occursin(restart_file_rx, f), readdir(folder_path))
+        if !isempty(possible_restart_files)
+            previous_folder = folder_path
+            restart_file_name = last(sort_files_by_time(possible_restart_files))
+            restart_file = joinpath(previous_folder, restart_file_name)
+            @assert isfile(restart_file) "Restart file does not exist"
+            return restart_file
+        end
+    end
+
+    # No folder with restart files found
+    @warn "No restart files found in any output folder in $(base_output_dir)"
+    return nothing
+end
+
+"""
+    setup_output_dir(job_id, output_dir, output_dir_style, detect_restart_file, restart_file, comms_ctx)
+
+Unified function for setting up output directories and detecting restart files.
+Used by both AtmosSimulation constructor and get_simulation.
+
+Returns a named tuple with:
+- `output_dir`: The final output directory path
+- `restart_file`: The restart file path (if any)
+"""
+function setup_output_dir(
+    job_id,
+    output_dir,
+    output_dir_style,
+    detect_restart_file,
+    restart_file,
+    comms_ctx,
+)
+    # Set up base output directory
+    default_output = haskey(ENV, "CI") ? job_id : joinpath("output", job_id)
+    base_output_dir = isnothing(output_dir) ? default_output : output_dir
+
+    # Validate and get output directory style
+    allowed_dir_styles = Dict(
+        "activelink" => OutputPathGenerator.ActiveLinkStyle(),
+        "removepreexisting" => OutputPathGenerator.RemovePreexistingStyle(),
+    )
+
+    haskey(allowed_dir_styles, lowercase(output_dir_style)) ||
+        error("output_dir_style $(output_dir_style) not available")
+
+    output_dir_style_obj = allowed_dir_styles[lowercase(output_dir_style)]
+
+    final_restart_file = if detect_restart_file && isnothing(restart_file)
+        auto_detect_restart_file(output_dir_style_obj, base_output_dir)
+    else
+        restart_file
+    end
+
+    # Generate the actual output directory
+    output_dir = OutputPathGenerator.generate_output_path(
+        base_output_dir;
+        context = comms_ctx,
+        style = output_dir_style_obj,
+    )
+
+    return output_dir, final_restart_file
+end
