+++ src/simulation/solve.jl
@@ -75,7 +75,7 @@ function solve_atmos!(simulation)
     precompile_callbacks(integrator)
     GC.gc()
     try
-        if CA.is_distributed(comms_ctx)
+        if is_distributed(comms_ctx)
             # GC.enable(false) # disabling GC causes a memory leak
             ClimaComms.barrier(comms_ctx)
             (sol, walltime) = timed_solve!(integrator)
@@ -87,11 +87,11 @@ function solve_atmos!(simulation)
             return AtmosSolveResults(sol, :success, walltime)
         end
     catch ret_code
-        if !CA.is_distributed(comms_ctx)
+        if !is_distributed(comms_ctx)
             # We can only save when not distributed because we don't have a way to sync the
             # MPI processes (maybe just one MPI rank crashes, leading to a hanging
             # simulation)
-            CA.save_state_to_disk_func(integrator, simulation.output_dir)
+            save_state_to_disk_func(integrator, simulation.output_dir)
         end
         @error "ClimaAtmos simulation crashed. Stacktrace for failed simulation" exception =
             (ret_code, catch_backtrace())
@@ -106,87 +106,10 @@ function solve_atmos!(simulation)
     end
 end
 
-"""
-    benchmark_step!(integrator)
-
-# Example
-
-To instantiate the environment:
-```
-julia --project=.buildkite
-```
-
-Then, to run interactively:
-```julia
-import ClimaAtmos as CA
-import Random
-Random.seed!(1234)
-(; config_file, job_id) = CA.commandline_kwargs();
-config = CA.AtmosConfig(config_file; job_id);
-simulation = CA.get_simulation(config);
-(; integrator) = simulation
-Y₀ = deepcopy(integrator.u);
-CA.benchmark_step!(integrator, Y₀);
-```
-
-Alternatively, this can be run non-interactively,
-with adjusted flags as:
-```
-julia --project=.buildkite perf/benchmark_step.jl --h_elem 6
-```
-
-See [`argparse_settings`](@ref) for the
-method defining the flags, and
-
-    https://clima.github.io/ClimaAtmos.jl/dev/cl_args/
-
-for the flags outlined in a table.
-
-"""
-function benchmark_step!(integrator, Y₀, n_steps = 10)
-    for i in 1:n_steps
-        CTS.step!(integrator)
-        integrator.u .= Y₀ # temporary hack to simplify performance benchmark.
-    end
-    return nothing
-end
-
-"""
-    cycle!(integrator; n_cycles = 1)
-
-Run `step!` the least common multiple times
-for all callbacks. i.e., all callbacks will
-have been called at least `n_cycles` times.
-
-`cycle!` is the true atomic unit for performance,
-because, unlike `step!`, it takes all callbacks
-into account.
-"""
-function cycle!(integrator; n_cycles = 1)
-    n_steps = n_steps_per_cycle(integrator) * n_cycles
-    for i in 1:n_steps
-        CTS.step!(integrator)
-    end
-    return nothing
-end
-
 function call_all_callbacks!(integrator)
     for cb! in atmos_callbacks(integrator.callback)
         cb!(integrator)
     end
-    return nothing
-end
-
-"""
-    precompile_atmos(integrator)
-
-Precompiles `step!` and all callbacks
-in the `integrator`.
-"""
-function precompile_atmos(integrator)
-    B = Base.precompile(CTS.step!, (typeof(integrator),))
-    @assert B
-    precompile_callbacks(integrator)
     return nothing
 end
 
@@ -195,7 +118,6 @@ function precompile_callbacks(integrator)
     @assert B
     return nothing
 end
-
 
 check_conservation(simulation::AtmosSimulation) =
     check_conservation(simulation.integrator.sol)
@@ -272,7 +194,6 @@ function write_diagnostics_as_txt(simulation::AtmosSimulation)
     )
     return nothing
 end
-
 
 """
     write_diagnostics_as_txt(writer, output_dir)
