+++ src/types.jl
@@ -1248,447 +1248,3 @@ function _throw_unknown_atmos_model_argument_error(unknown_args)
         join(available_all, "\n  "),
     )
 end
-
-abstract type AbstractCallbackFrequency end
-struct EveryNSteps <: AbstractCallbackFrequency
-    n::Int
-end
-struct EveryΔt{FT} <: AbstractCallbackFrequency
-    Δt::FT
-end
-struct AtmosCallback{F, CBF <: AbstractCallbackFrequency, VI <: Vector{Int}}
-    f!::F
-    cbf::CBF
-    n_measured_calls::VI
-end
-callback_frequency(cb::AtmosCallback) = cb.cbf
-prescribed_every_n_steps(x::EveryNSteps) = x.n
-prescribed_every_n_steps(cb::AtmosCallback) = prescribed_every_n_steps(cb.cbf)
-
-prescribed_every_Δt_steps(x::EveryΔt) = x.Δt
-prescribed_every_Δt_steps(cb::AtmosCallback) = prescribed_every_Δt_steps(cb.cbf)
-
-# TODO: improve accuracy
-n_expected_calls(cbf::EveryΔt, dt, tspan) = (tspan[2] - tspan[1]) / cbf.Δt
-n_expected_calls(cbf::EveryNSteps, dt, tspan) =
-    ((tspan[2] - tspan[1]) / dt) / cbf.n
-n_expected_calls(cb::AtmosCallback, dt, tspan) =
-    n_expected_calls(cb.cbf, dt, tspan)
-
-AtmosCallback(f!, cbf) = AtmosCallback(f!, cbf, Int[0])
-function (cb::AtmosCallback)(integrator)
-    cb.f!(integrator)
-    cb.n_measured_calls[] += 1
-    return nothing
-end
-n_measured_calls(cb::AtmosCallback) = cb.n_measured_calls[]
-
-struct AtmosConfig{FT, TD, PA, C, CF}
-    toml_dict::TD
-    parsed_args::PA
-    comms_ctx::C
-    config_files::CF
-    job_id::String
-end
-
-Base.eltype(::AtmosConfig{FT}) where {FT} = FT
-
-TupleOrVector(T) = Union{Tuple{<:T, Vararg{T}}, Vector{<:T}}
-
-# Use short, relative paths, if possible.
-function normrelpath(file)
-    rfile = normpath(relpath(file, dirname(config_path)))
-    return if isfile(rfile) && samefile(rfile, file)
-        rfile
-    else
-        file
-    end
-end
-
-function maybe_add_default(config_files, default_config_file)
-    return if any(x -> samefile(x, default_config_file), config_files)
-        config_files
-    else
-        (default_config_file, config_files...)
-    end
-end
-
-"""
-    AtmosConfig(
-        config_file::String = default_config_file;
-        job_id = config_id_from_config_file(config_file),
-        comms_ctx = nothing,
-    )
-    AtmosConfig(
-        config_files::Union{NTuple{<:Any, String} ,Vector{String}};
-        job_id = config_id_from_config_files(config_files),
-        comms_ctx = nothing,
-    )
-
-Helper function for the AtmosConfig constructor. Reads a YAML file into a Dict
-and passes it to the AtmosConfig constructor.
-"""
-AtmosConfig(
-    config_file::String = default_config_file;
-    job_id = config_id_from_config_file(config_file),
-    comms_ctx = nothing,
-) = AtmosConfig((config_file,); job_id, comms_ctx)
-
-function AtmosConfig(
-    config_files::TupleOrVector(String);
-    job_id = config_id_from_config_files(config_files),
-    comms_ctx = nothing,
-)
-
-    all_config_files =
-        normrelpath.(maybe_add_default(config_files, default_config_file))
-    configs = map(all_config_files) do config_file
-        strip_help_messages(load_yaml_file(config_file))
-    end
-    return AtmosConfig(
-        configs;
-        comms_ctx,
-        config_files = all_config_files,
-        job_id,
-    )
-end
-
-"""
-    AtmosConfig(
-        configs::Union{NTuple{<:Any, Dict} ,Vector{Dict}};
-        comms_ctx = nothing,
-        config_files,
-        job_id
-    )
-
-Constructs the AtmosConfig from the Dicts passed in. This Dict overrides all of
-the default configurations set in `default_config_dict()`.
-"""
-AtmosConfig(configs::AbstractDict; kwargs...) =
-    AtmosConfig((configs,); kwargs...)
-function AtmosConfig(
-    configs::TupleOrVector(AbstractDict);
-    comms_ctx = nothing,
-    config_files = [default_config_file],
-    job_id = "",
-)
-    config_files = map(x -> normrelpath(x), config_files)
-
-    # using config_files = [default_config_file] as a default
-    # relies on the fact that override_default_config uses
-    # default_config_file.
-    config = merge(configs...)
-    comms_ctx = isnothing(comms_ctx) ? get_comms_context(config) : comms_ctx
-    config = override_default_config(config)
-
-    FT = config["FLOAT_TYPE"] == "Float64" ? Float64 : Float32
-    atmos_toml = map(config["toml"]) do file
-        isfile(file) ? file :
-        isfile(joinpath(pkgdir(@__MODULE__), file)) ?
-        joinpath(pkgdir(@__MODULE__), file) : error("Parameter file $file not found.")
-    end
-    toml_dict = CP.create_toml_dict(
-        FT;
-        override_file = CP.merge_toml_files(atmos_toml),
-    )
-    config = config_with_resolved_and_acquired_artifacts(config, comms_ctx)
-
-    isempty(job_id) &&
-        @warn "`job_id` is empty and likely not passed to AtmosConfig"
-
-    @info "Making AtmosConfig with config files: $(sprint(config_summary, config_files))"
-
-    C = typeof(comms_ctx)
-    TD = typeof(toml_dict)
-    PA = typeof(config)
-    CF = typeof(config_files)
-    return AtmosConfig{FT, TD, PA, C, CF}(
-        toml_dict,
-        config,
-        comms_ctx,
-        config_files,
-        job_id,
-    )
-end
-
-# AtmosConfig-aware constructors for the AtmosModel group structs.
-# Each consolidates the YAML→typed-object translation for one group.
-
-function AtmosWater(config::AtmosConfig, params, ::Type{FT}) where {FT}
-    pa = config.parsed_args
-    microphys_model_str = pa["microphysics_model"]
-    @assert microphys_model_str in ("dry", "0M", "1M", "2M", "2MP3")
-    microphysics_model =
-        if microphys_model_str == "dry"
-            DryModel()
-        elseif microphys_model_str == "0M"
-            EquilibriumMicrophysics0M()
-        elseif microphys_model_str == "1M"
-            NonEquilibriumMicrophysics1M(;
-                n_substeps = pa["microphysics_n_substeps"],
-                n_substeps_quad = pa["microphysics_n_substeps_quadrature"],
-            )
-        elseif microphys_model_str == "2M"
-            NonEquilibriumMicrophysics2M()
-        elseif microphys_model_str == "2MP3"
-            NonEquilibriumMicrophysics2MP3()
-        end
-
-    sgs_quadrature = get_sgs_quadrature(pa, params)
-
-    if microphysics_model isa DryModel
-        @warn "Running simulations without any moisture present."
-    end
-    if microphysics_model isa EquilibriumMicrophysics0M && isnothing(sgs_quadrature)
-        error(
-            "EquilibriumMicrophysics0M requires use_sgs_quadrature: true. " *
-            "GridMeanSGS fallback is not supported for 0-moment microphysics.",
-        )
-    end
-
-    cloud_model = get_cloud_model(pa, params)
-
-    terminal_velocity_mode =
-        pa["fixed_terminal_velocity"] ?
-        FixedTerminalVelocity{FT}(
-            CAP.fixed_cloud_liquid_terminal_velocity(params),
-            CAP.fixed_cloud_ice_terminal_velocity(params),
-            CAP.fixed_rain_terminal_velocity(params),
-            CAP.fixed_snow_terminal_velocity(params),
-        ) : DiagnosticTerminalVelocity()
-
-    implicit_microphysics = pa["implicit_microphysics"]
-    @assert implicit_microphysics in (true, false)
-
-    return AtmosWater(;
-        microphysics_model,
-        cloud_model,
-        microphysics_tendency_timestepping = implicit_microphysics ? Implicit() :
-                                             Explicit(),
-        tracer_nonnegativity_method = get_tracer_nonnegativity_method(pa),
-        sgs_quadrature,
-        terminal_velocity_mode,
-    )
-end
-
-function AtmosRadiation(config::AtmosConfig, ::Type{FT}; setup_type = nothing) where {FT}
-    pa = config.parsed_args
-
-    radiation_mode = get_radiation_mode(pa, FT; setup_type)
-
-    # Legacy `forcing: held_suarez` selects HeldSuarezForcing as the radiation
-    # mode (deprecated; prefer `rad: held_suarez`).
-    forcing = pa["forcing"]
-    @assert forcing in (nothing, "held_suarez")
-    if forcing == "held_suarez"
-        @warn "The 'held_suarez' forcing option is deprecated. Use rad='held_suarez' instead."
-        radiation_mode = HeldSuarezForcing()
-    end
-
-    return AtmosRadiation(;
-        radiation_mode,
-        insolation = get_insolation_form(pa; setup_type),
-    )
-end
-
-function AtmosGravityWave(config::AtmosConfig, params, ::Type{FT}) where {FT}
-    pa = config.parsed_args
-    return AtmosGravityWave(;
-        non_orographic_gravity_wave = get_non_orographic_gravity_wave_model(pa, params, FT),
-        orographic_gravity_wave = get_orographic_gravity_wave_model(pa, params, FT),
-    )
-end
-
-function AtmosTurbconv(config::AtmosConfig, params, ::Type{FT}) where {FT}
-    pa = config.parsed_args
-    turbconv_params = CAP.turbconv_params(params)
-
-    implicit_sgs_advection = pa["implicit_sgs_advection"]
-    @assert implicit_sgs_advection in (true, false)
-    implicit_sgs_entr_detr = pa["implicit_sgs_entr_detr"]
-    @assert implicit_sgs_entr_detr in (true, false)
-    implicit_sgs_nh_pressure = pa["implicit_sgs_nh_pressure"]
-    @assert implicit_sgs_nh_pressure in (true, false)
-    implicit_sgs_vertdiff = pa["implicit_sgs_vertdiff"]
-    @assert implicit_sgs_vertdiff in (true, false)
-    implicit_sgs_mass_flux = pa["implicit_sgs_mass_flux"]
-    @assert implicit_sgs_mass_flux in (true, false)
-
-    scale_blending_method =
-        if pa["edmfx_scale_blending"] == "SmoothMinimum"
-            SmoothMinimumBlending()
-        elseif pa["edmfx_scale_blending"] == "HardMinimum"
-            HardMinimumBlending()
-        else
-            error("Unknown edmfx_scale_blending method: $(pa["edmfx_scale_blending"])")
-        end
-
-    edmfx_model = EDMFXModel(;
-        entr_model = get_entrainment_model(pa),
-        detr_model = get_detrainment_model(pa),
-        sgs_mass_flux = pa["edmfx_sgs_mass_flux"],
-        sgs_diffusive_flux = pa["edmfx_sgs_diffusive_flux"],
-        nh_pressure = pa["edmfx_nh_pressure"],
-        vertical_diffusion = pa["edmfx_vertical_diffusion"],
-        filter = pa["edmfx_filter"],
-        scale_blending_method,
-    )
-
-    n = pa["smagorinsky_lilly"]
-    smagorinsky_lilly =
-        isnothing(n) ? nothing : SmagorinskyLilly(; axes = Symbol(n))
-
-    amd_les_active = pa["amd_les"]
-    @assert amd_les_active in (true, false)
-    amd_les = amd_les_active ? AnisotropicMinimumDissipation{FT}(pa["c_amd"]) : nothing
-
-    chd_active = pa["constant_horizontal_diffusion"]
-    @assert chd_active in (true, false)
-    constant_horizontal_diffusion =
-        chd_active ?
-        ConstantHorizontalDiffusion{FT}(CAP.constant_horizontal_diffusion_D(params)) :
-        nothing
-
-    return AtmosTurbconv(;
-        edmfx_model,
-        turbconv_model = get_turbconv_model(FT, pa, turbconv_params),
-        sgs_adv_mode = implicit_sgs_advection ? Implicit() : Explicit(),
-        sgs_entr_detr_mode = implicit_sgs_entr_detr ? Implicit() : Explicit(),
-        sgs_nh_pressure_mode = implicit_sgs_nh_pressure ? Implicit() : Explicit(),
-        sgs_vertdiff_mode = implicit_sgs_vertdiff ? Implicit() : Explicit(),
-        sgs_mf_mode = implicit_sgs_mass_flux ? Implicit() : Explicit(),
-        smagorinsky_lilly,
-        amd_les,
-        constant_horizontal_diffusion,
-    )
-end
-
-AtmosNumerics(config::AtmosConfig, ::Type{FT}) where {FT} =
-    get_numerics(config.parsed_args, FT)
-
-function SCMSetup(config::AtmosConfig, ::Type{FT};
-    setup_type = nothing) where {FT}
-    pa = config.parsed_args
-
-    advection_test = pa["advection_test"]
-    @assert advection_test in (false, true)
-
-    return SCMSetup(;
-        subsidence = get_subsidence_model(FT; setup_type),
-        external_forcing = get_external_forcing_model(pa, FT; setup_type),
-        ls_adv = get_large_scale_advection_model(FT; setup_type),
-        advection_test,
-        scm_coriolis = get_scm_coriolis(FT; setup_type),
-    )
-end
-
-function AtmosSponge(config::AtmosConfig, params)
-    pa = config.parsed_args
-
-    vs_name = pa["viscous_sponge"]
-    viscous_sponge = if vs_name in ("false", false, "none")
-        nothing
-    elseif vs_name in ("true", true, "ViscousSponge")
-        ViscousSponge(params)
-    else
-        error("Uncaught viscous sponge model `$vs_name`.")
-    end
-
-    rs_name = pa["rayleigh_sponge"]
-    rayleigh_sponge = if rs_name in ("false", false)
-        nothing
-    elseif rs_name in ("true", true, "RayleighSponge")
-        RayleighSponge(params)
-    else
-        error("Uncaught rayleigh sponge model `$rs_name`.")
-    end
-
-    return AtmosSponge(; viscous_sponge, rayleigh_sponge)
-end
-
-function AtmosSurface(
-    config::AtmosConfig, params, ::Type{FT}; setup_type = nothing,
-) where {FT}
-    pa = config.parsed_args
-
-    sfc_temperature = Setups.surface_temperature_model(setup_type)
-
-    surface_model =
-        if pa["prognostic_surface"] in ("false", false, "PrescribedSST")
-            PrescribedSST()
-        elseif pa["prognostic_surface"] in ("true", true, "SlabOceanSST")
-            SlabOceanSST()
-        else
-            error("Uncaught surface model `$(pa["prognostic_surface"])`.")
-        end
-
-    surface_albedo =
-        if pa["albedo_model"] == "ConstantAlbedo"
-            ConstantAlbedo{FT}(; α = params.idealized_ocean_albedo)
-        elseif pa["albedo_model"] == "RegressionFunctionAlbedo"
-            isnothing(pa["rad"]) && error(
-                "Radiation model not specified, so cannot use RegressionFunctionAlbedo",
-            )
-            RegressionFunctionAlbedo{FT}(; n = params.water_refractive_index)
-        elseif pa["albedo_model"] == "CouplerAlbedo"
-            CouplerAlbedo()
-        else
-            error("Uncaught surface albedo model `$(pa["albedo_model"])`.")
-        end
-
-    return AtmosSurface(; sfc_temperature, surface_model, surface_albedo)
-end
-
-"""
-    maybe_resolve_and_acquire_artifacts(input_str::AbstractString, context)
-
-When given a string of the form `artifact"name"/something/else`, resolve the
-artifact path and download it (if not already available).
-
-In all the other cases, return the input unchanged.
-"""
-function maybe_resolve_and_acquire_artifacts(
-    input_str::AbstractString,
-    context,
-)
-    matched = match(r"artifact\"([a-zA-Z0-9_]+)\"(\/.*)?", input_str)
-    if isnothing(matched)
-        return input_str
-    else
-        artifact_name, other_path = matched
-        return joinpath(
-            @clima_artifact(string(artifact_name), context),
-            lstrip(other_path, '/'),
-        )
-    end
-end
-
-function maybe_resolve_and_acquire_artifacts(
-    input,
-    _,
-)
-    return input
-end
-
-"""
-    config_with_resolved_and_acquired_artifacts(input_str::AbstractString, context)
-
-Substitute strings of the form `artifact"name"/something/else` with the actual
-artifact path.
-"""
-function config_with_resolved_and_acquired_artifacts(
-    config::AbstractDict,
-    context,
-)
-    return Dict(
-        k => maybe_resolve_and_acquire_artifacts(v, context) for
-        (k, v) in config
-    )
-end
-
-function config_summary(io::IO, config_files)
-    print(io, '\n')
-    for x in config_files
-        println(io, "   $x")
-    end
-end
